+++ src/components/Item.tsx
@@ -8,7 +8,9 @@ import {
   ItemName,
   ItemPrice,
 } from '@/styles/Item/Item.styles';
-import { useProductDetail } from '@/hooks/product';
+import { useNavigate } from 'react-router-dom';
+import { useContext } from 'react';
+import { LoginInfoContext } from '@/contexts/LoginInfoContext';
 
 type ItemProps = {
   index: number;
