+++ src/styles/emotion.d.ts
@@ -1,5 +1,5 @@
 import '@emotion/react';
-import { theme } from '@/styles/ResetStyles';
+import { theme } from '@/styles/theme';
 
 type AppTheme = typeof theme;
